Shuffled functions in core.py; added/fixed tests; other cleanup
This is a squashed commit from a local branch created a
few days ago.  I added several debugging tools for
myself, and I spent a lot of time revisiting the code I
wrote over the past few months.  There was little
refactoring and renaming, but the main changes are the
new tests, new debugging tools, and reordering the
functions in core.py.