Added get_Copeland_winners() function, and prep for fixing IRV support
My work over the past week has been a mish-mash of all sorts of cleanup
and debugging, and I got exhausted.  The previous version did a better
job of handling weird characters inside of square brackets or inside of
double quotes, but this version fails with a candidate token of
"[Doña #1]" or "Steve #1" or whatever.  One of these days, I'll fix that
problem, but not today.