Tweaked behavior of --add-scores, fixing scorestar_test.py test
When scores/ratings get added, the ratings that get added should
depend on whether the input file has ratings.  If it has ratings for
some entries, then assume a default of zero for the entries that don't
have ratings.  If there are no ratings throughout the input file, then
assume the ratings to be added are Borda-ish.