Fix(cli): Fix cli when no parameters given (+test)
This regression was introduced in 25083b4 when adding the --container
option (for archives). This is the first time it was valid NOT to have
an inputfile. This bug made it clear there weren't any tests for basic
CLI functionality, so there's now at least "cli_test.py" which
currently tests for no parameters, and is a good home for other CLI
tests in the future.