test(core): Add tests for trailing space in ABIF files
mock-scorestar-tie.abif accidentally found a trailing space problem in
the ABIF parser. I want to fix that eventually, but I'd like to fix
the score tie issue too (mabye before fixing the parser errors).  It
would seem that I have all sorts of under-the-hood unimplemented
features with score and JSON as well.